+++ src/components/atoms/ResultsAmount/ResultsAmount.test.tsx
