+++ src/components/atoms/ResultsAmount/ResultsAmount.tsx
