+++ src/components/atoms/index.ts
@@ -1,4 +1,4 @@
-export * from "./BackButton";
-export * from "./Button";
-export * from "./Input";
-export * from "./ResultsAmount";
+export * from "./BackButton/BackButton";
+export * from "./Button/Button";
+export * from "./Input/Input";
+export * from "./ResultsAmount/ResultsAmount";
