+++ src/components/organisms/index.ts
@@ -1,1 +1,1 @@
-export * from "./Navbar";
+export * from "./Navbar/Navbar";
